Changed major place.cpp data structures from file scope to global scope. Moved delay routines to place_delay_model.*. Moved annealing update routines to place_util.*. Enhanced documentations.
